refactor(cron): split timer watchdog regression ownership
